requirements.txt
Django==3.1.6
djangorestframework==3.12.2

numpy~=1.20.1
model-bakery==1.2.1
coverage==5.4
